% Generated by build script
\name{npm_package}
\alias{npm_package}
\title{Get npm package metadata}
\usage{
npm_package()
}
\arguments{
\item{name}{Package name (e.g. "express", "react", "lodash")}
}
\value{tibble: name, version, description, license, homepage,}
\description{Get npm package metadata}
