% Generated by build script
\name{npm_search}
\alias{npm_search}
\title{Search npm registry for packages}
\usage{
npm_search()
}
\arguments{
\item{text}{Search query string}
\item{size}{Number of results to return (default 20, max 250)}
}
\value{tibble: name, version, description, keywords, date, author,}
\description{Search npm registry for packages}
